requirements.txt
mysql-connector-python==8.0.30
python-dotenv==1.0.0
colorlog==6.7.0